pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "swagger-mcp"
version = "0.1.0"
description = "MCP server for OpenAPI/Swagger document analysis, supporting OpenAPI 3.0 and Swagger 2.0"
authors = [
    {name = "Vibe Coding", email = ""}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastmcp>=0.1.0",
    "pydantic>=2.0.0",
    "openapi-spec-validator>=0.7.0",
    "requests>=2.31.0",
    "pyyaml>=6.0",
    "typing-extensions>=4.8.0",
]

[project.scripts]
swagger-mcp = "swagger_mcp.server:run_server" 